--- src/model_meta/training_config.yaml
@@ -1,27 +1,0 @@
-hydra:
-  run:
-    dir: "./logs/${now:%Y-%m%d-%H%M-%S}/"
-  
-#data_path: "/home/takeru/AlphaSymbol/data/training/mini.csv"
-#metadata_path: "/home/takeru/AlphaSymbol/data/training/mini_metadata.yaml"
-data_path: "/Users/takeruito/work/PrfSR/data/training/superfib_r1_dataset.csv"
-metadata_path: "/Users/takeruito/work/PrfSR/data/training/superfib_r1_metadata.yaml"
-max_epoch: 1
-max_value: 2000 # max value occurd in input and output of PRE
-min_n_tokens_in_batch: 2000
-test_ratio: 0.1 # (test + val) / (test + val + train)
-val_ratio: 0.25 # val / (val + test)
-num_workers: 13
-token_embed_dim: 16
-emb_expansion_factor: 1
-learning_rate: "3*10**(-4)"
-batching_strategy: "length_aware_token" # "dynamic" or "fixed"
-batch_size: 64
-
-transformer:
-  nhead: 16
-  num_encoder_layers: 4
-  num_decoder_layers: 6
-  dim_feedforward: 512
-  dropout: 0.1
-
